moving exercise to global events, adding convertible note valuation tracking
